PKCS11: Fix implementation of PKCS#11 2.11 paragraph 11.4
According to the specification, if CKF_OS_LOCKING_OK is set and 'fields
are supplied (i.e., they all have nonNULL_PTR values)' the library can
decide to use app locking or os locking (locking with 'the native
operating system primitives') or return CKR_CANT_LOCK (which is what it
currently does). Since the library is already using system primitives to
implement locking, it is de-facto using os locking and it should not
return CKR_CANT_LOCK because it can actually lock.

This is critical for allowing the library to be loaded by p11-kit in a
manged way which in turn allows it to be loaded by p11-kit-proxy.

Also:
    - Comment unused string.
    - Throw CKR_CRYPTOKI_ALREADY_INITIALIZED if the library is already
      initialized.

Note related to p11-kit:
In the scenario where p11-kit loads both opensc-pkcs11 and libcie-pkcs11
the Smart Card reader is picked up by both modules and it is likely that
both modules will try to access the CIE. This means opensc is going to
get stuck at reading the CIE (because it cannot read it properly). As a
result the user will be unable to use the CIE.
As a workaround the user should tell opensc to ignore the smart card
reader in opensc settings (i.e. adding 'ignored_readers = Reader Name;'
to the proper section of opensc.conf)

Signed-off-by: Luca Magrone <luca@magrone.cc>