CIEID: Override turorial to explain the OpenSC issue
Replace the remote webpage loaded from the tutorial tab with a bundled
one that reflects the new changes introduced with the set of patches
included in the RPM package.

Specifically, inform the user that no further configuration has to be
done unless there is OpenSC installed on the system. Walk the user
through what to do if that's the case.

Signed-off-by: Luca Magrone <luca@magrone.cc>