PKCS11: Fix implementation of PKCS#11 2.11 paragraph 11.4
According to the specification, if CKF_OS_LOCKING_OK is set and 'fields are supplied (i.e. they all have nonNULL_PTR values)' the library can decide to use app locking (using the given primitives) or os locking (ignoring the primitives and locking with 'the native operating system primitives') or, if it can't lock, return CKR_CANT_LOCK (which is what it currently does). Since the library is already using system primitives to implement locking, it is already using os locking and it should not return CKR_CANT_LOCK because it can actually lock. That's why it doesn't throw an error if the caller of the library doesn't set function pointers.

Signed-off-by: Luca Magrone <luca@magrone.cc>