cie_sign_sdk: PdfSignatureGenerator: Rewrite implementation of PADES signature
As of now the code relies on a closed-source version of PoDoFo.
Write a new implementation of the PdfSignatureGenerator class based on
the code from podofosign and the old code of the class. Try to match
behaviour to the behaviour of versions for other platforms.
Use open-source PoDoFo.
Also fix loading document for verification with the correct method.

Signed-off-by: Luca Magrone <luca@magrone.cc>