CSP: firmaConCIE: fix error path
Currently the code tries to make a signature with every token in every
card reader. This way if there are multiple tokens (either CIEs or not)
the signature can be completed but an error pops up to the user wich is bad
UX. Furthermore if an unrecognised token or a wrong CIE is evaluated
before the correct CIE the function fails, which should not.

Keep looking up for the correct CIE until all card readers are evaluated.
Do not immediately fail if an unrecognised token is found.
Fail on PAN mismatch only if the mismatch happens on every CIE.
Stop looking if the correct CIE is found.

Also free allocated resources before returning.

Signed-off-by: Luca Magrone <luca@magrone.cc>